Give the line descriptor cookie its own type (`ldesc`)
The line descriptor was an `int` cookie, which encoded a file
offset pointer and a mark flag, with bit-twiddling and masking
to manipulate it. This had knock-on effects in `blkio.c`.

Making the descriptor a `struct` with two fields removes some of the
complexity in `blkio.c`, and makes the Qed address logic clearer.